requirements.txt
discord-py
youtube-search-python
python-dotenv
pytube